% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{available_discretization_methods}
\alias{available_discretization_methods}
\title{Available discretization methods}
\usage{
available_discretization_methods()
}
\value{
Character vector of method names.
}
\description{
Available discretization methods
}
