% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_epsilon_net}
\alias{cluster_epsilon_net}
\title{Epsilon-net clustering}
\usage{
cluster_epsilon_net(W, epsilon = 3)
}
\arguments{
\item{W}{Symmetric binary adjacency matrix.}

\item{epsilon}{Radius parameter. Only \code{epsilon = 3} is supported.}
}
\value{
A list with entry \code{clusters}.
}
\description{
Epsilon-net clustering
}
