% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_louvain_membership}
\alias{cluster_louvain_membership}
\title{Louvain community detection}
\usage{
cluster_louvain_membership(W)
}
\arguments{
\item{W}{Symmetric binary adjacency matrix.}
}
\value{
Integer cluster-membership vector.
}
\description{
Louvain community detection
}
