% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{graph_laplacian}
\alias{graph_laplacian}
\title{Compute a graph Laplacian}
\usage{
graph_laplacian(W, normalized = TRUE, binary = TRUE)
}
\arguments{
\item{W}{Symmetric matrix used to construct the graph Laplacian.}

\item{normalized}{If \code{TRUE}, return the normalized Laplacian.}

\item{binary}{If \code{TRUE}, use the binary support of \code{W}.}
}
\value{
A square matrix.
}
\description{
Compute a graph Laplacian
}
