% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cluster_count_by_sample_size}
\alias{plot_cluster_count_by_sample_size}
\title{Plot number of clusters by sample size}
\usage{
plot_cluster_count_by_sample_size(extracted_results)
}
\arguments{
\item{extracted_results}{Data frame produced by an extraction workflow over
repeated network simulations.}
}
\value{
A list of ggplot objects, one per network type.
}
\description{
Plot number of clusters by sample size
}
