% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{causalClust-package}
\alias{causalClust}
\alias{causalClust-package}
\title{causalClust: causal clustering for network experiments}
\description{
Tools for constructing and evaluating cluster randomized designs under
network interference. The main design functions are
\code{\link[=causal_clustering_algorithm1]{causal_clustering_algorithm1()}} and \code{\link[=causal_clustering_algorithm2]{causal_clustering_algorithm2()}}.
}
\author{
\strong{Maintainer}: Oleksandr Stasovskyi \email{ostasovskyi@college.harvard.edu} [contributor]

Authors:
\itemize{
  \item Davide Viviano
  \item Lihua Lei
  \item Guido Imbens
  \item Brian Karrer
  \item Okke Schrijvers
  \item Liang Shi
}

}
\keyword{internal}
